migrate to commons-jexl3, add quarkus-poi instead of poi (simplifies management) , upgrade poi to 5.4.1

+++ src/main/java/net/sf/jett/model/FontCache.java
@@ -51,7 +51,7 @@ public class FontCache
      */
     private void cachePreExistingFonts()
     {
-        short numFonts = myWorkbook.getNumberOfFonts();
+       int numFonts = myWorkbook.getNumberOfFonts();
         logger.trace("Caching {} pre-existing cell fonts.", numFonts);
         for (short i = 0; i < numFonts; i++)
         {
