+++ src/test/java/net/sf/jett/test/model/AreaCellListener.java
@@ -70,7 +70,7 @@ public class AreaCellListener implements CellListener
                 Workbook workbook = cell.getSheet().getWorkbook();
                 CellStyle style = workbook.createCellStyle();
                 style.cloneStyleFrom(cell.getCellStyle());
-                short fontIdx = style.getFontIndex();
+                int fontIdx = style.getFontIndex();
                 Font font = workbook.getFontAt(fontIdx);
                 Font italicFont = workbook.findFont(font.getBold(), font.getColor(), font.getFontHeight(),
                         font.getFontName(), true, font.getStrikeout(), font.getTypeOffset(),
