+++ src/test/java/net/sf/jett/test/model/ItalicTagListener.java
@@ -53,7 +53,7 @@ public class ItalicTagListener implements TagListener
                     {
                         Workbook workbook = sheet.getWorkbook();
                         CellStyle style = cell.getCellStyle();
-                        short fontIdx = style.getFontIndex();
+                        int fontIdx = style.getFontIndex();
                         Font font = workbook.getFontAt(fontIdx);
                         if (!font.getItalic())
                         {
